Resolution of the bugs

+++ src/app/modules/usuarios/components/modal/modal.component.ts
@@ -3,6 +3,7 @@ import {UserListResponse} from "../../../../core/models/user.model";
 import {take} from "rxjs/operators";
 import {UsuariosService} from "../../services/usuarios.service";
 import {ToastrService} from "ngx-toastr";
+import {AuthService} from "../../../auth/services/auth.service";
 
 
 @Component({
@@ -24,7 +25,8 @@ export class ModalComponent implements AfterViewInit {
   @Input() public txtButton: string;
 
   constructor(private _userService: UsuariosService,
-              private _toastrService: ToastrService) {
+              private _toastrService: ToastrService,
+              private _authService: AuthService) {
     this.onCloseModal = new EventEmitter<boolean>();
     this.title = '';
     this.user = {id: 0, name: '', lastName: '', email: '', username:'', password:'', image:'', phone: 0}
@@ -45,16 +47,18 @@ export class ModalComponent implements AfterViewInit {
 
   public deleteUsertById(): void {
     if (this.txtButton === 'Eliminar') {
+      if (this.user.email !== this._authService.getAuthToken()) {
       this._userService.deleteUserById(this.user.id)
         .pipe(take(1))
         .subscribe(() => {
-          this._toastrService.warning(`${this.user.name} eliminado con éxito`, 'Eliminar')
+          this._toastrService.warning(`${this.user.name} fue eliminado exitosamente`, 'Eliminar')
         }, (error) => {
-          this._toastrService.error(`Ocurrió un error al eliminar el usuario`)
+          this._toastrService.error(`${this.user.name} no fue eliminado, intente nuevamente`)
         });
+      }else {
+        this._toastrService.error(`No es posible eliminar tu propia cuenta`, 'Error')
+      }
     }
     this.closeModal();
   }
-
-
 }
